// src/types/city.enum.ts
export enum CityType {
  ALL = 'Вся Україна',
  KYIV = 'м. Київ',
  VINNYTSKA = 'Вінницька',
  VOLYNSKA = 'Волинська',
  DNIPROPETROVSKA = 'Дніпропетровська',
  DONETSKA = 'Донецька',
  ZHYTOMYRSKA = 'Житомирська',
  ZAKARPATSKA = 'Закарпатська',
  ZAPORIZKA = 'Запорізька',
  IVANOFRANKIVSKA = 'Івано-Франківська',
  KYIVSKA = 'Київська',
  KIROVOHRADSKA = 'Кіровоградська',
  LUHANSKA = 'Луганська',
  LVIVSKA = 'Львівська',
  MYKOLAIVSKA = 'Миколаївська',
  ODESSKA = 'Одеська',
  POLTAVSKA = 'Полтавська',
  RIVNENSKA = 'Рівненська',
  SUMSKA = 'Сумська',
  TERNOPILSKA = 'Тернопільська',
  KHARKIVSKA = 'Харківська',
  KHERSONSKA = 'Херсонська',
  KHMELNYTSKA = 'Хмельницька',
  CHERKASKA = 'Черкаська',
  CHERNIVETSKA = 'Чернівецька',
  CHERNIGIVSKA = 'Чернігівська',
}